added product CRUD

+++ database/migrations/2022_03_11_163139_create_products_table.php
@@ -16,7 +16,7 @@ return new class extends Migration {
             $table->id();
             $table->string('name');
             $table->text('description');
-            $table->unsignedBigInteger('price');
+            $table->double('price');
             $table->unsignedBigInteger('stock');
             $table->unsignedBigInteger('category_id');
             $table->foreign('category_id')->references('id')->on('categories')->onDelete('cascade');
